# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "4ks recipe API.\n\nAuthentication uses `Authorization: Bearer \u003ctoken\u003e`.\nMost authenticated routes expect an Auth0 JWT.\nApproved recipe routes documented as AI Kitchen Pass compatible also accept a Kitchen Pass bearer token.",
        "title": "4ks API",
        "contact": {},
        "version": "2.0"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Recipes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get All Recipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Recipe"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Recipe. This route accepts either an Auth0 JWT or an AI Kitchen Pass bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new Recipe",
                "parameters": [
                    {
                        "description": "Recipe Data",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateRecipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                }
            }
        },
        "/api/recipes/author/{username}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Recipes by Author",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get All Recipes by Author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetRecipesByUsernameResponse"
                        }
                    }
                }
            }
        },
        "/api/recipes/fetch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Request Recipe Fetch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Request Recipe Fetch",
                "parameters": [
                    {
                        "description": "Recipe Data",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.FetchRecipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateUserEvent"
                        }
                    }
                }
            }
        },
...
        "/api/recipes/revisions/{revisionID}/fork": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fork a specific historical recipe revision. This route accepts either an Auth0 JWT or an AI Kitchen Pass bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Fork Recipe Revision",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Revision ID",
                        "name": "revisionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                }
            }
        },
        "/api/recipes/{recipeID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a Recipe (by ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get a Recipe (by ID)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.GetRecipeResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Recipe. This route accepts either an Auth0 JWT or an AI Kitchen Pass bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Update Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe Data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateRecipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                }
            }
        },
...
        "/api/recipes/{recipeID}/forks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get direct child forks for a Recipe. This route is part of the AI Kitchen Pass workflow surface.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get direct forks for a Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Recipe"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.CreateRecipe": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeMediaVariant"
                    }
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instruction"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.CreateUserEvent": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "$ref": "#/definitions/models.UserEventStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.UserEventType"
                }
            }
        },
        "dtos.FetchRecipeRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dtos.GetRecipeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Recipe"
                }
            }
        },
        "dtos.GetRecipesByUsernameResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Recipe"
                    }
                }
            }
        },
...
        "dtos.UpdateRecipe": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeMediaVariant"
                    }
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instruction"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Ingredient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Instruction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.Recipe": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "branch": {
                    "type": "string"
                },
                "contributors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserSummary"
                    }
                },
                "createdDate": {
                    "type": "string"
                },
                "currentRevision": {
                    "$ref": "#/definitions/models.RecipeRevision"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.RecipeMetadata"
                },
                "root": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        },
...
        "models.RecipeMediaVariant": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "maxWidth": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.RecipeMetadata": {
            "type": "object",
            "properties": {
                "forks": {
                    "type": "integer"
                },
                "stars": {
                    "type": "integer"
                }
            }
        },
        "models.RecipeRevision": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "banner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeMediaVariant"
                    }
                },
                "createdDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instruction"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recipeId": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        },
...
        "models.UserEventStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                9,
                60
            ],
            "x-enum-varnames": [
                "UserEventCreated",
                "UserEventProcessing",
                "UserEventReady",
                "UserEventAcknowledged",
                "UserEventExpired",
                "UserEventErrorState"
            ]
        },
        "models.UserEventType": {
            "type": "integer",
            "enum": [
                0,
                2,
                3,
                9
            ],
            "x-enum-varnames": [
                "UserEventTypeNewUser",
                "UserEventTypeCreateRecipe",
                "UserEventTypeForkRecipe",
                "UserEventTypeFetchRecipe"
            ]
        },
        "models.UserSummary": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}